Improve signed_in? verification
We currently determine if the user is `signed_in?` by checking that the `ExtensionAuthorizer` has been configured with an account. This however, is problematic because the constructor ensures that the `@user` instance variable _always_ [has a value](https://github.com/sensu/bonsai/blob/master/app/lib/authorizer/base.rb#L16).

Ultimately this gap led to a lot of spam emails to our administrator accounts.

--- 

In terms of determining whether a user is able to report an asset we may want to do more than check that the user is "signed in". For instance, we may want to check if they've linked their Github account or otherwise has confirmed their email. And / or include a captcha.